fix bash installation srcipt #2

+++ install.sh
@@ -2,12 +2,14 @@
 
 # Vim
 
-runtime_path=~\.vim_runtime
-vimrc_source=${runtime_path}\vimrc
-vimrc_destination=~\.vimrc
-gvimrc_destination=~\.gvimrc
+runtime_path=~/.vim_runtime
+vimrc_source=$runtime_path/vimrc
+vimrc_destination=~/.vimrc
+gvimrc_destination=~/.gvimrc
 
-ln -s $vimrc_source $vimrc_destination
+if [[ ! -f $vimrc_destination ]]; then
+    ln -s $vimrc_source $vimrc_destination
+fi
 touch $gvimrc_destination
 
 # Neovim
@@ -18,10 +20,14 @@ if [[ ! -d $nvim_config_dir ]]; then
     mkdir -p $nvim_config_dir
 fi
 
-nvimrc_destination=${nvim_config_dir}\init.vim
-gnvimrc_source=${runtime_path}\ginit.vim
-gnvimrc_destination=${nvim_config_dir}\ginit.vim
+nvimrc_destination=$nvim_config_dir/init.vim
+gnvimrc_source=$runtime_path/ginit.vim
+gnvimrc_destination=$nvim_config_dir/ginit.vim
 
-ln -s $vimrc_source $nvimrc_destination
-ln -s $gnvimrc_source $gnvimrc_destination
+if [[ ! -f $nvimrc_destination ]]; then
+    ln -s $vimrc_source $nvimrc_destination
+fi
+if [[ ! -f $gnvimrc_destination ]]; then
+    ln -s $gnvimrc_source $gnvimrc_destination
+fi
